move getCoursesList and updateCartItems function to utils
